Add files via upload
This program downloads the open data of COVID-19 from the website of the Ministry of Health, Labour and Welfare, the Government of Japan, and computes the numerical solutions to the initial value problem for the SIR model. I am very sorry to say that the number of deaths are included in the active cases in order to compare with the SIR model. I would like to express his sincere condolence of the victims and their family members.